# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Allyourbase API
  description: |
    Backend-as-a-Service for PostgreSQL. Single binary. Auto-generated REST API
    with filtering, sorting, pagination, foreign key expansion, full-text search,
    batch operations, authentication, file storage, realtime events, and webhooks.
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Allyourbase
    url: https://allyourbase.io
servers:
  - url: http://localhost:8090
    description: Local development server

# Default: no auth required. Individual operations override with BearerAuth or AdminAuth.
security:
  - {}
...
paths:
...
  /api/admin/api-keys/{id}:
    delete:
      tags: [Admin API Keys]
      summary: Revoke an API key
      operationId: adminRevokeApiKey
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: API key revoked
        "400":
          description: Invalid UUID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/storage/{bucket}:
    get:
      tags: [Storage]
      summary: List files in a bucket
      operationId: listFiles
      parameters:
        - $ref: "#/components/parameters/BucketPath"
        - name: prefix
          in: query
          description: Filter by filename prefix
          schema:
            type: string
        - name: limit
          in: query
          description: Max files to return
          schema:
            type: integer
        - name: offset
          in: query
          description: Number of files to skip
          schema:
            type: integer
      security:
        - BearerAuth: []
        - {}
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageListResponse"
        "400":
          description: Invalid bucket name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Storage]
      summary: Upload a file
      operationId: uploadFile
      parameters:
        - $ref: "#/components/parameters/BucketPath"
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
              required:
                - file
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageObject"
        "400":
          description: Invalid form or file too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/storage/{bucket}/{name}:
    get:
      tags: [Storage]
      summary: Download a file
      operationId: downloadFile
      parameters:
        - $ref: "#/components/parameters/BucketPath"
        - $ref: "#/components/parameters/FileName"
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Storage]
      summary: Delete a file
      operationId: deleteFile
      parameters:
        - $ref: "#/components/parameters/BucketPath"
        - $ref: "#/components/parameters/FileName"
      security:
        - BearerAuth: []
        - {}
      responses:
        "204":
          description: File deleted
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/storage/{bucket}/{name}/sign:
    post:
      tags: [Storage]
      summary: Get a signed URL for a file
      description: Generate a time-limited signed URL. Default expiry is 1 hour, max 7 days.
      operationId: signFileUrl
      parameters:
        - $ref: "#/components/parameters/BucketPath"
        - $ref: "#/components/parameters/FileName"
      security:
        - BearerAuth: []
        - {}
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignRequest"
      responses:
        "200":
          description: Signed URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignResponse"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/auth/login or /api/auth/register
    AdminAuth:
      type: http
      scheme: bearer
      description: Admin token from /api/admin/auth
  parameters:
...
    BucketPath:
      name: bucket
      in: path
      required: true
      description: Storage bucket name
      schema:
        type: string
    FileName:
      name: name
      in: path
      required: true
      description: File name within the bucket
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "collection not found: nonexistent"
        data:
          type: object
          description: Field-level validation details
          additionalProperties: true
...
    StorageObject:
      type: object
      required: [id, bucket, name, size, contentType, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        bucket:
          type: string
        name:
          type: string
        size:
          type: integer
          format: int64
          description: File size in bytes
        contentType:
          type: string
          example: image/png
        userId:
          type: string
          format: uuid
          description: Uploader's user ID (if authenticated)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StorageListResponse:
      type: object
      required: [items, totalItems]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StorageObject"
        totalItems:
          type: integer
    SignRequest:
      type: object
      properties:
        expiresIn:
          type: integer
          description: Expiry in seconds (default 3600, max 604800)
          default: 3600
          maximum: 604800
    SignResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Signed URL path with token query parameter
...